for anaasyab domain

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -13,27 +12,29 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('created_by')->nullable()->constrained('users')->onDelete('cascade');
-            $table->enum('gender' ,['male' , 'female'])->nullable();
-            $table->string('name');
-            $table->string('level')->default('user');
-            $table->string('admin_token')->nullable()->unique();
-            $table->string('lastname');
-            $table->string('email')->nullable()->unique();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->string('phone', 11)->unique();
-            $table->timestamp('phone_verified_at')->nullable();
-            $table->string('password')->nullable();
-            $table->string('province')->nullable();
-            $table->string('city')->nullable();
-            $table->string('slug')->nullable();
-            $table->boolean('active')->default(true);
-            $table->rememberToken();
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (env('REFRESH_MIGRATION')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('created_by')->nullable()->constrained('users')->onDelete('cascade');
+                $table->enum('gender', ['male', 'female'])->nullable();
+                $table->string('name');
+                $table->string('level')->default('user');
+                $table->string('admin_token')->nullable()->unique();
+                $table->string('lastname');
+                $table->string('email')->nullable()->unique();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->string('phone', 11)->unique();
+                $table->timestamp('phone_verified_at')->nullable();
+                $table->string('password')->nullable();
+                $table->string('province')->nullable();
+                $table->string('city')->nullable();
+                $table->string('slug')->nullable();
+                $table->boolean('active')->default(true);
+                $table->rememberToken();
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -43,6 +44,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+        if (env('REFRESH_MIGRATION')) {
+            Schema::dropIfExists('users');
+        }
     }
 };
