+++ database/migrations/2023_04_16_092635_create_provinces_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -13,14 +12,16 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('provinces', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('country_id')->constrained('countries')->onDelete('cascade');
-            $table->string('name');
-            $table->string('name_en')->nullable();
-            $table->decimal('latitude' , 10 ,8)->nullable();
-            $table->decimal('longitude' , 10 ,8)->nullable();
-        });
+        if (env('REFRESH_MIGRATION')) {
+            Schema::create('provinces', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('country_id')->constrained('countries')->onDelete('cascade');
+                $table->string('name');
+                $table->string('name_en')->nullable();
+                $table->decimal('latitude', 10, 8)->nullable();
+                $table->decimal('longitude', 10, 8)->nullable();
+            });
+        }
     }
 
     /**
@@ -30,6 +31,7 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('provinces');
+        if (env('REFRESH_MIGRATION')) {
+            Schema::dropIfExists('provinces');}
     }
 };
