+++ database/migrations/2023_04_16_092717_create_cities_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -13,15 +12,16 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('cities', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('province_id')->constrained('provinces')->onDelete('cascade');
-            $table->string('name');
-            $table->string('name_en')->nullable();
-            $table->decimal('latitude' , 10 ,8)->nullable();
-            $table->decimal('longitude' , 10 ,8)->nullable();
-        });
-
+        if (env('REFRESH_MIGRATION')) {
+            Schema::create('cities', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('province_id')->constrained('provinces')->onDelete('cascade');
+                $table->string('name');
+                $table->string('name_en')->nullable();
+                $table->decimal('latitude', 10, 8)->nullable();
+                $table->decimal('longitude', 10, 8)->nullable();
+            });
+        }
 //        DB::table('users')->insert(
 //            array(
 //                'email' => 'name@domain.example',
@@ -37,6 +37,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('cities');
+        if (env('REFRESH_MIGRATION')) {
+            Schema::dropIfExists('cities');
+        }
     }
 };
