+++ database/migrations/2023_04_16_092753_create_workshops_table.php
@@ -13,26 +13,28 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('workshops', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('creator_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('city_id')->constrained('cities')->onDelete('cascade');
-            $table->string('title');
-            $table->string('description')->nullable();
-            $table->text('body')->nullable();
-            $table->string('slug')->nullable();
-            $table->timestamp('event_time');
-            $table->integer('capacity')->default(0);
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (env('REFRESH_MIGRATION')) {
+            Schema::create('workshops', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('creator_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('city_id')->constrained('cities')->onDelete('cascade');
+                $table->string('title');
+                $table->string('description')->nullable();
+                $table->text('body')->nullable();
+                $table->string('slug')->nullable();
+                $table->timestamp('event_time');
+                $table->integer('capacity')->default(0);
+                $table->softDeletes();
+                $table->timestamps();
+            });
 
-        Schema::create('category_workshop', function (Blueprint $table) {
-            $table->foreignId('workshop_id')->constrained('workshops')->onDelete('cascade');
-            $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
-            $table->primary(['workshop_id', 'category_id']);
-            $table->timestamps();
-        });
+            Schema::create('category_workshop', function (Blueprint $table) {
+                $table->foreignId('workshop_id')->constrained('workshops')->onDelete('cascade');
+                $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
+                $table->primary(['workshop_id', 'category_id']);
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -42,8 +44,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('category_workshop');
+        if (env('REFRESH_MIGRATION')) {
+            Schema::dropIfExists('category_workshop');
 
-        Schema::dropIfExists('workshops');
+        Schema::dropIfExists('workshops');}
     }
 };
