+++ database/migrations/2023_04_16_092849_create_tickets_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -13,19 +12,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('tickets', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('creator_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
-            $table->foreignId('workshop_id')->constrained('workshops')->onDelete('cascade');
-            $table->string('name');
-            $table->string('last_name');
-            $table->string('phone');
-            $table->string('email')->nullable();
-            $table->text('token')->unique();
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (env('REFRESH_MIGRATION')) {
+            Schema::create('tickets', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('creator_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
+                $table->foreignId('workshop_id')->constrained('workshops')->onDelete('cascade');
+                $table->text('token')->unique();
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -35,6 +33,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('tickets');
+        if (env('REFRESH_MIGRATION')) {
+            Schema::dropIfExists('tickets');
+        }
     }
 };
