+++ database/migrations/2023_04_16_092945_create_galleries_table.php
@@ -13,14 +13,15 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('galleries', function (Blueprint $table) {
+        if (env('REFRESH_MIGRATION')) {
+            Schema::create('galleries', function (Blueprint $table) {
             $table->id();
             $table->foreignId('creator_id')->constrained('users')->onDelete('cascade');
             $table->string('title')->nullable();
             $table->morphs('galleryable');
             $table->softDeletes();
             $table->timestamps();
-        });
+        });}
     }
 
     /**
@@ -30,6 +31,7 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('galleries');
+        if (env('REFRESH_MIGRATION')) {
+            Schema::dropIfExists('galleries');}
     }
 };
