+++ database/migrations/2023_04_16_093156_create_blogs_table.php
@@ -13,25 +13,26 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('blogs', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('creator_id')->constrained('users')->onDelete('cascade');
-            $table->string('title');
-            $table->string('description');
-            $table->text('body')->nullable();
-            $table->string('slug')->nullable();
-            $table->integer('viewCount')->default(0);
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (env('REFRESH_MIGRATION')) {
+            Schema::create('blogs', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('creator_id')->constrained('users')->onDelete('cascade');
+                $table->string('title');
+                $table->string('description');
+                $table->text('body')->nullable();
+                $table->string('slug')->nullable();
+                $table->integer('viewCount')->default(0);
+                $table->softDeletes();
+                $table->timestamps();
+            });
 
-        Schema::create('blog_category', function (Blueprint $table) {
-            $table->foreignId('blog_id')->constrained('blogs')->onDelete('cascade');
-            $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
-            $table->primary(['blog_id', 'category_id']);
-            $table->timestamps();
-        });
-
+            Schema::create('blog_category', function (Blueprint $table) {
+                $table->foreignId('blog_id')->constrained('blogs')->onDelete('cascade');
+                $table->foreignId('category_id')->constrained('categories')->onDelete('cascade');
+                $table->primary(['blog_id', 'category_id']);
+                $table->timestamps();
+            });
+        }
     }
 
 
@@ -43,7 +44,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('blog_category');
-        Schema::dropIfExists('blogs');
+        if (env('REFRESH_MIGRATION')) {
+            Schema::dropIfExists('blog_category');
+        Schema::dropIfExists('blogs');}
     }
 };
