+++ database/migrations/2023_04_16_113007_create_announcements_table.php
@@ -13,7 +13,8 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('announcements', function (Blueprint $table) {
+        if (env('REFRESH_MIGRATION')) {
+            Schema::create('announcements', function (Blueprint $table) {
             $table->id();
             $table->foreignId('workshop_id')->nullable()->constrained('workshops')->onDelete('cascade');
 //            $table->foreignId('classroom_id')->nullable()->constrained('classrooms')->onDelete('cascade');
@@ -26,7 +27,7 @@ return new class extends Migration
             $table->string('status')->default('pending');
             $table->softDeletes();
             $table->timestamps();
-        });
+        });}
     }
 
     /**
@@ -36,6 +37,7 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('announcements');
+        if (env('REFRESH_MIGRATION')) {
+            Schema::dropIfExists('announcements');}
     }
 };
