+++ database/migrations/2023_04_16_113435_create_payments_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      *
@@ -13,16 +12,18 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('payments', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('creator_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
-            $table->string('transaction_id');
-            $table->string('price');
-            $table->boolean('payment')->default(false);
-            $table->softDeletes();
-            $table->timestamps();
-        });
+        if (env('REFRESH_MIGRATION')) {
+            Schema::create('payments', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('creator_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
+                $table->string('transaction_id');
+                $table->string('price');
+                $table->boolean('payment')->default(false);
+                $table->softDeletes();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -32,6 +33,8 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('payments');
+        if (env('REFRESH_MIGRATION')) {
+            Schema::dropIfExists('payments');
+        }
     }
 };
